elevenlabsdemo/requirements.txt
elevenlabs>=2.0.0
python-dotenv>=1.0.0
pyaudio>=0.2.13
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
